Refactor: move TradeChatOffer from chat/pub/ to chat/
