+++ desktop/src/main/java/bisq/desktop/primary/main/content/components/ChatMessagesListView.java
@@ -28,7 +28,7 @@ import bisq.chat.trade.priv.PrivateTradeChatMessage;
 import bisq.chat.trade.pub.PublicTradeChannel;
 import bisq.chat.trade.pub.PublicTradeChannelService;
 import bisq.chat.trade.pub.PublicTradeChatMessage;
-import bisq.chat.trade.pub.TradeChatOffer;
+import bisq.chat.trade.TradeChatOffer;
 import bisq.common.monetary.Coin;
 import bisq.common.monetary.Fiat;
 import bisq.common.observable.Pin;
