+++ desktop/src/main/java/bisq/desktop/primary/overlay/createOffer/review/ReviewOfferView.java
@@ -18,7 +18,7 @@
 package bisq.desktop.primary.overlay.createOffer.review;
 
 import bisq.chat.trade.pub.PublicTradeChatMessage;
-import bisq.chat.trade.pub.TradeChatOffer;
+import bisq.chat.trade.TradeChatOffer;
 import bisq.common.currency.Market;
 import bisq.common.monetary.Fiat;
 import bisq.desktop.common.utils.Transitions;
